+++ lib/creq.rb
@@ -4,11 +4,15 @@ require_relative "creq/requirement"
 require_relative "creq/parser"
 require_relative "creq/reader"
 require_relative "creq/repository"
+require_relative "creq/req_cmd"
+require_relative "creq/settings"
 require_relative "helper"
 require_relative "cli"
 
 module Creq
   include Helper
+
+  CreqCmdError = Class.new(StandardError)
 
   def self.root
     File.dirname __dir__