+++ lib/creq/settings.rb
@@ -1,0 +1,13 @@
+# encoding: UTF-8
+
+module Creq
+  module Settings
+    DOC = 'doc'
+    LIB = 'lib'
+    REQ = 'req'
+    TT = 'tt'
+    DOC_ASSETS = 'doc/assets'
+
+    PROJECT_FOLDERS = [DOC, LIB, REQ, TT, DOC_ASSETS]
+  end
+end